TIP-1285: Rename subscribers
