+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/event_subscribers.yml
@@ -55,15 +55,15 @@ services:
         tags:
             - { name: doctrine.event_subscriber, priority: 50 }
 
-    pim_catalog.event_subscriber.delete_products_from index:
-        class: 'Akeneo\Pim\Enrichment\Bundle\EventSubscriber\DeleteProductsFromIndexSubscriber'
+    pim_catalog.event_subscriber.product.delete.compute_products_and_ancestors:
+        class: 'Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Product\OnDelete\ComputeProductsAndAncestorsSubscriber'
         arguments:
             - '@akeneo.pim.enrichment.elasticsearch.indexer.product_and_ancestors'
         tags:
             - { name: kernel.event_subscriber }
 
-    pim_catalog.event_subscriber.compute_completeness_and_index_products:
-        class: 'Akeneo\Pim\Enrichment\Bundle\EventSubscriber\ComputeCompletenessAndIndexProductsSubscriber'
+    pim_catalog.event_subscriber.product.save.compute_products_and_ancestors:
+        class: 'Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Product\OnSave\ComputeProductsAndAncestorsSubscriber'
         arguments:
             - '@pim_catalog.completeness.product.compute_and_persist'
             - '@akeneo.pim.enrichment.elasticsearch.indexer.product_and_ancestors'
